osa 4 part 4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -33,8 +33,28 @@ blogsRouter.post('/', async (request, response) => {
   response.status(201).json(savedBlog)
 })
 
-// 4.18 seuraavaksi
+
 blogsRouter.delete('/:id', async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  // check token
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  // check if blog exists
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({ error: 'blog with this id could not be found' })
+  }
+  // check that the blog that is being deleted is the users own blog
+  const user = await User.findById(decodedToken.id)
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'Wrong user' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
